requirements.txt
django
djangorestframework
streamlit
google-generativeai
requests